4.21*: bloglist expansion, step9

+++ part4/blog-list/controllers/blogs.js
@@ -34,9 +34,30 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const id = request.params.id
-  await Blog.findByIdAndRemove(id)
-  response.status(204).end()
+  const { id } = request.params
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  // check for valid token
+  if (!request.token || !decodedToken || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(id)
+
+  const user = await User.findById(decodedToken.id)
+
+  // Check if creator of blog is user trying to delete it.
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(id)
+    response.status(204).end()
+  } else {
+    response
+      .status(401)
+      .json({ error: 'You are not authorized to delete this blog' })
+  }
+
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
